+++ apps/docs/registry/registry-ui-components/drawer/drawer.types.ts
@@ -1,0 +1,28 @@
+import { UseDuckAlertReturnType } from '../alert-dialog/alert-dialog.types'
+import {
+  DrawerClose,
+  DrawerContent,
+  DrawerDescription,
+  DrawerFooter,
+  DrawerHeader,
+  DrawerTitle,
+  DrawerTrigger,
+} from './drawer'
+
+/**
+ * DrawerWrapperProps
+ */
+export type DrawerWrapperProps = {
+  trigger?: React.ComponentPropsWithoutRef<typeof DrawerTrigger>
+  content: Partial<React.ComponentPropsWithoutRef<typeof DrawerContent>> & {
+    _header?: React.ComponentPropsWithoutRef<typeof DrawerHeader> & {
+      _title?: React.ComponentPropsWithoutRef<typeof DrawerTitle>
+      _description?: React.ComponentPropsWithoutRef<typeof DrawerDescription>
+    }
+    _footer?: React.ComponentPropsWithoutRef<typeof DrawerFooter> & {
+      _cancel?: React.ComponentPropsWithoutRef<typeof DrawerClose>
+      _submit?: React.HTMLProps<HTMLDivElement>
+    }
+  }
+  duckHook?: UseDuckAlertReturnType
+}
