+++ apps/docs/registry/registry-ui-components/drawer/index.ts
@@ -1,0 +1,2 @@
+export * from './drawer'
+export * from './drawer.types'
